+++ src/main/java/com/youthcase/orderflow/sd/sdSales/service/SDService.java
@@ -40,5 +40,5 @@ public interface SDService {
 
     List<SalesHeaderDTO> getHoldOrders();
 
-    void updateItemQuantity(Long itemId, int quantity);
+    void updateItemQuantity(Long itemId, Long quantity);
 }
