+++ src/main/java/com/youthcase/orderflow/stk/dto/AdjustmentRequest.java
@@ -13,6 +13,6 @@ public class AdjustmentRequest {
     public static class AdjustmentItem {
         private Long lotId;           // 조정할 재고의 랏 ID
         private String productGtin;   // 제품 GTIN (선택 사항)
-        private int targetQuantity;   // ⭐️ 조정하고자 하는 최종 수량 (폐기에서는 'quantity'였으나, 조정에서는 목표 수량)
+        private Long targetQuantity;   // ⭐️ 조정하고자 하는 최종 수량 (폐기에서는 'quantity'였으나, 조정에서는 목표 수량)
     }
 }
