+++ src/main/java/com/youthcase/orderflow/stk/dto/DisposalRequest.java
@@ -12,6 +12,6 @@ public class DisposalRequest {
     public static class DisposalItem {
         private Long lotId;           // 폐기할 재고의 랏 ID (Lot 테이블의 PK 또는 STK 테이블이 참조하는 ID)
         private String productGtin;   // 제품 GTIN (필수는 아니지만 검증에 유용)
-        private int quantity;         // 요청된 폐기 수량
+        private Long quantity;         // 요청된 폐기 수량
     }
 }
